adding networking config

+++ vital_site/vital/utils.py
@@ -38,21 +38,33 @@ class XenClient:
         xen = LoadBalancer().get_best_server()
         logger.debug(len(course.virtual_machine_set.all()))
         for vm in course.virtual_machine_set.all():
-            available_config = Available_Config.objects.filter(category='MAC_ADDR').order_by('id')[0]
-            locked_conf = Available_Config.objects.select_for_update().filter(id=available_config.id)[0]
+            flag = True
+            cnt = 0
+            # hack to handle concurrent requests
+            while flag:
+                available_config = Available_Config.objects.filter(category='MAC_ADDR').order_by('id')[0]
+                locked_conf = Available_Config.objects.select_for_update().filter(id=available_config.id)[0]
+                cnt += 1
+                if locked_conf is not None:
+                    val = locked_conf.value
+                    locked_conf.delete()
+                    # TODO change this to accept other private networks
+                    # Done just to accept class nets
+                    class_net = vm.network_configuration_set.filter(is_course_net=True)[0]
+                    vif = '\'mac=' + val + ', bridge=' + class_net.name + '\''
+                    xen.setup_vm(user, str(user.id) + '_' + str(course.id) + '_' + str(vm.id),
+                                 str(course.id) + '_' + str(vm.id), vif)
+                    user_net_config = User_Network_Configuration()
+                    user_net_config.bridge_name = class_net.name
+                    user_net_config.user_id = user.id
+                    user_net_config.mac_id = val
+                    user_net_config.vm = vm
+                    user_net_config.save()
+                    flag = False
+                if cnt >= 100:
+                    raise Exception('Server Busy : Registration incomplete')
 
-            # TODO change this to accept other private networks
-            # Done just to accept class nets
-            class_net = vm.network_configuration_set.filter(is_course_net=True)[0]
-            vif = '\'mac='+locked_conf.value+', bridge='+class_net.name+'\''
-            xen.setup_vm(user, str(user.id)+'_'+str(course.id)+'_'+str(vm.id), str(course.id)+'_'+str(vm.id), vif)
-            user_net_config = User_Network_Configuration()
-            user_net_config.bridge_name=class_net.name
-            user_net_config.user_id = user.id
-            user_net_config.mac_id = locked_conf.value
-            user_net_config.vm = vm
-            user_net_config.save()
-            locked_conf.delete()
+
 
 
     def unregister_student_vms(self, server, user, course):
